development/tke-privacy-engine/functions/deid-processor/requirements.txt
functions-framework>=3.0.0
google-genai>=1.0.0
google-cloud-dlp>=3.0.0
google-cloud-firestore>=2.0.0
google-cloud-storage>=2.0.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
PyPDF2>=3.0.0
python-docx>=1.0.0
pyyaml>=6.0.0
pydantic>=2.10.0
